LayoutTests/fast/canvas/fillrect_gradient-expected.txt

Each square canvas should contain a colored gradient bordered by a narrow white margin and a black line. The column on the left contains linear gradients, the column on the right radial gradients.
 Green to white
 Green to white to red
 Larger green to white to larger red
 Solid red
 Solid red
 Solid red
 Blue to white in the top (inner) half, red to yellow in the bottom (outer) half
 Blue to white, red to yellow (same as previous)
 Blue to white, red to yellow (same as previous)
